+++ react-movie-catalog-app/src/components/MoviesPreview.tsx
@@ -4,7 +4,8 @@ import { MoviePreview } from './MoviePreview';
 import { movieType } from './type';
 import { MovieOperation } from './MovieOperation';
 import { MoviesContext } from '../context/MoviesContext';
-import { TMovie, TMoviesPreviewProps } from '../ts-types/types';
+import { TMovie } from '../ts-types/movie';
+import { TMoviesPreviewProps } from '../ts-types/props';
 
 export const MoviesPreview = ({ addMovie, viewMovie }: TMoviesPreviewProps) => {
   const [addModalWindow, setAddModalWindow] = useState(false);
