+++ react-movie-catalog-app/src/components/Notification.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import successImg from './../assets/img/notifications/success/success.png';
 import errorImg from './../assets/img/notifications/error/error.png';
-import { TNotificationProps } from '../ts-types/types';
+import { TNotificationProps } from '../ts-types/props';
 
 export const Notification = ({ type, message, description, onClose }: TNotificationProps) => {
   return (
