+++ react-movie-catalog-app/src/hooks/useMovie.ts
@@ -1,5 +1,5 @@
 import * as NotFoundImg from '../assets/img/404/404.png';
-import { TMovie } from '../ts-types/types';
+import { TMovie } from '../ts-types/movie';
 export const useMovie = (movie: TMovie) => {
   const { poster_path = NotFoundImg } = movie;
   return {
